+++ packages/apidom-ls/src/config/openapi/header/lint/index.ts
@@ -1,0 +1,29 @@
+import allowedFields3_0Lint from './allowed-fields-3-0';
+import allowedFields3_1Lint from './allowed-fields-3-1';
+import descriptionTypeLint from './description--type';
+import requiredTypeLint from './required--type';
+import deprecatedTypeLint from './deprecated--type';
+import allowEmptyValueTypeLint from './allow-empty-value--type';
+import styleTypeLint from './style--type';
+import explodeTypeLint from './explode--type';
+import allowReservedTypeLint from './allow-reserved--type';
+import schemaTypeLint from './schema--type';
+import examplesValuesTypeLint from './examples--values-type';
+import contentValuesTypeLint from './content--values-type';
+
+const lints = [
+  descriptionTypeLint,
+  requiredTypeLint,
+  deprecatedTypeLint,
+  allowEmptyValueTypeLint,
+  styleTypeLint,
+  explodeTypeLint,
+  allowReservedTypeLint,
+  schemaTypeLint,
+  examplesValuesTypeLint,
+  contentValuesTypeLint,
+  allowedFields3_0Lint,
+  allowedFields3_1Lint,
+];
+
+export default lints;