+++ packages/apidom-ls/src/config/openapi/header/lint/style--type.ts
@@ -1,0 +1,16 @@
+import ApilintCodes from '../../../codes';
+import { LinterMeta } from '../../../../apidom-language-types';
+
+const styleTypeLint: LinterMeta = {
+  code: ApilintCodes.OPENAPI3_0_HEADER_FIELD_STYLE_TYPE,
+  source: 'apilint',
+  message: 'style must be a string',
+  severity: 1,
+  linterFunction: 'apilintType',
+  linterParams: ['string'],
+  marker: 'value',
+  target: 'style',
+  data: {},
+};
+
+export default styleTypeLint;