+++ packages/apidom-ls/src/config/openapi/paths/lint/keys--pattern.ts
@@ -4,10 +4,10 @@ import { LinterMeta } from '../../../../apidom-language-types';
 const keysPatternLint: LinterMeta = {
   code: ApilintCodes.OPENAPI3_0_PATHS_KEYS_PATTERN,
   source: 'apilint',
-  message: 'Paths Object keys must match the regular expression: `^/`',
+  message: 'Paths Object keys must match the regular expression: `^(/|x-)`',
   severity: 1,
   linterFunction: 'apilintKeysRegex',
-  linterParams: ['^/'],
+  linterParams: ['^(/|x-)'],
   marker: 'key',
   data: {},
 };