+++ packages/apidom-ls/src/config/openapi/request-body/lint/allowed-fields-3-0.ts
@@ -1,7 +1,8 @@
 import ApilintCodes from '../../../codes';
 import { LinterMeta } from '../../../../apidom-language-types';
 
-const allowedFieldsLint: LinterMeta = {
+// eslint-disable-next-line @typescript-eslint/naming-convention
+const allowedFields3_0Lint: LinterMeta = {
   code: ApilintCodes.NOT_ALLOWED_FIELDS,
   source: 'apilint',
   message: 'Object includes not allowed fields',
@@ -9,6 +10,12 @@ const allowedFieldsLint: LinterMeta = {
   linterFunction: 'allowedFields',
   linterParams: [['description', 'content', 'required', '$ref'], 'x-'],
   marker: 'key',
+  targetSpecs: [
+    { namespace: 'openapi', version: '3.0.0' },
+    { namespace: 'openapi', version: '3.0.1' },
+    { namespace: 'openapi', version: '3.0.2' },
+    { namespace: 'openapi', version: '3.0.3' },
+  ],
 };
 
-export default allowedFieldsLint;
+export default allowedFields3_0Lint;