+++ packages/apidom-ls/src/config/openapi/request-body/lint/index.ts
@@ -1,19 +1,15 @@
-import allowedFieldsLint from './allowed-fields';
-import $refNoSiblingsLint from './$ref--no-siblings';
-import $refValidLint from './$ref--valid';
+import allowedFields3_0Lint from './allowed-fields-3-0';
 import descriptionTypeLint from './description--type';
 import contentValuesTypeLint from './content--values-type';
 import contentRequiredLint from './content--required';
 import requiredTypeLint from './required--type';
 
 const lints = [
-  allowedFieldsLint,
-  $refNoSiblingsLint,
-  $refValidLint,
   descriptionTypeLint,
   contentRequiredLint,
   contentValuesTypeLint,
   requiredTypeLint,
+  allowedFields3_0Lint,
 ];
 
 export default lints;